requirements.txt
nose
coverage
pysqlite
flask-sqlalchemy
